Refactor classnames and addPoint function to take ID
In this commit, I passed a string to identify what type of player was having their points added, as opposed to passing the element through.
I like passing `this` through as an argument, it's quite elegant and shows a good understanding of javascript.
However, in line 36 of app.js, you then have to use javascript to fetch another element anyway, so if you're doing that, you might as well do that from the start.
Also, fetching nextElementSibling can be quite brittle. What if you want to introduce another element in the parent div? The functionality might break.

I've also changed the addPoint function to pass in an ID. At the moment it is just home and away, but this can be extended to be any player ID. It just means that if we choose to display multiple games (like at the world championships) we can. It becomes a bit more extendible. Each player should have a unique ID which we can use to identify, instead of home and away.

I've changed index.html to use more classes than ID's. This allows for a bit more extendibility. If I need it to be a bit more specific, I just add another class name and target it with two class names. I try and keep styling and functional class names separate, but this never quite works out sometimes...

At this point, I think there are more lines of code, but that's OK. Changing code around can make it balloon up a bit, until you've finished the entire thought process. I like to commit when I've made a small change to the code base and the functionality still works. That way if I really F it up, then I can go back. I might have the last commit message say that this change is complete.

Changing code while keeping its functionality is called Refactoring. There is a tonne of material about refactoring, mostly around Object Oriented Programming (OOP). There is a rather large camp against OOP as well, worth taking a look at.