Disable addPoint when game is won
I've added a bit of functionality to disable the addPoint function once
the match is over.

I also changed the for loop to use a forEach which is a function you get
for free when you use querySelectorAll.

Also, I think I might have introduced a bug, the game count should be 3
not 2. Not sure where that came from...