Create methods for calculating game and match state
This one isn't that much change, but the concept behind it takes a bit more explanation.

An object is often described as consisting of behaviour, that is coupled with data. So if we have a set of data that we want to compute to produce some results, that qualifies as a candidate for object transformation.

Here we have a Player that looks like This
Player {
    name,
    points,
    games,
}

This means that any behaviour (i.e. functions that use that data) we need, we can stick on the object.
The app cares about whether a player has won a game so it can update the DOM. This means that it doesn't care about HOW a game is calculated, or what rules are involved in working out whether a player has won a game. It just cares about if that player has won, yes or no.

Yes or no questions can be contained in a convenient method (a function that belongs to an object) that the app can ask the player object.

App -----> Have you completed a game yet? -----> Player
DOM <----- Update the DOM with new game score <----- App <------ Yes <------ Player

I'll draw better diagrams when I'm not on a bus without internet XD

This commit is all about that change. First I look at what logic can be contained inside a function. Then I see if that function is extracting data from an object, and whether or not I can stick that function on the object. I also question whether conceptually that makes sense (this is a lot more subjective obviously). Can I ask a player whether they have won the game or not? As a human being that would be silly because I know how a game works, I have the knowledge to calculate that game so I simply ask for the data. But machine objects should be made as simply as possible so they don't explode in complexity. This is what we're trying to tend towards, ultimately avoiding complexity until ABSOLUTELY necessary. Not trying to hammer a nail with a grenade and all that.

This is probably the newest concept so feel free to make comments on the commit and ask questions. In fact I demand you ask questions!