Refactor document interactions into score class
This change moves all logic that updates the DOM into an object. I
happened to call the object score this time, sometimes I'll call it a
type of View object, i.e. ScoreView. Views tend to contain the logic
needed to update whatever the User is going to view, just a convention
of naming.

By writing functions to contain the logic needed to update the DOM, you
can start to see what bits of code are duplicated, and what bits are
unique, i.e. what is constant, and what is variable.

The selector used to grab the pointsCount and the gamesCount are
largely the same, except for the playerId and the differentiation
between points and games.

.score.home > .pointsCount
.score.home > .gamesCount

So I extracted bits that are the same into a template, and the bits that
are variable into arguments:

countSelector(currentPlayerId, scoreType) {
  return '.score.' + currentPlayerId + ' > .' + scoreType
}

I carried on doing this for all DOM interactions and voila, the Score
object was born.